clean up configurations, default mmol_to_counts
